Rename fusion files and move them to separate folders
