+++ src/sensor/fusion/xiofusion/xioFusionOffset2.c
@@ -8,7 +8,7 @@
 //------------------------------------------------------------------------------
 // Includes
 
-#include "FusionOffset2.h"
+#include "xioFusionOffset2.h"
 #include <math.h> // fabsf
 
 //------------------------------------------------------------------------------
